Graphs added & Made Exp Page Component wise
