--- src/ml_board/frontend/Dashboard/src/components/dashboard/graphs/LineGraph.module.css
@@ -1,6 +1,0 @@
-.graphs_chart_container {
-    position: relative;
-    margin: auto;
-    height: 80vh;
-    width: 100%;
-}
