+++ src/ml_board/frontend/Dashboard/src/components/experimentPage/ExperimentPage.tsx
@@ -1,0 +1,169 @@
+import { Toolbar, styled, AccordionProps, AccordionSummaryProps, Card, CardContent, Box, Grid } from '@mui/material';
+import { useEffect, useState } from 'react';
+import { useSearchParams } from 'react-router-dom';
+import { useAppSelector } from "../../app/hooks";
+import { selectRowById } from '../../redux/table/tableSlice';
+import ArrowForwardIosSharpIcon from '@mui/icons-material/ArrowForwardIosSharp';
+import MuiAccordion from '@mui/material/Accordion';
+import MuiAccordionSummary from '@mui/material/AccordionSummary';
+import MuiAccordionDetails from '@mui/material/AccordionDetails';
+import styles from './ExperimentPage.module.css';
+import styles_graphs from "../graphs/Graphs.module.css";
+import Graph from "../graphs/Graph";
+import { getRandomColor } from '../../worker_socket/event_handlers/evaluationResultDataHandler';
+import CheckpointConfigurations from './CheckpointConfigurations/CheckpointConfigurations';
+import ExperimentConfigurations from './ExperimentConfigurations/ExperimentConfigurations';
+import { isConnected } from "../../redux/status/statusSlice";
+import { ExperimentDetails } from './ExperimentDetails.tsx/ExperimentDetails';
+import { ExperimentProgress } from './ExperimentProgress/ExperimentProgress';
+import { selectChartsByExperimentId } from '../../redux/charts/chartsSlice';
+
+export interface AnyKeyValuePairsInterface {
+    [key: string]: any
+}
+var anyObj: AnyKeyValuePairsInterface = {}
+
+interface lossOrMetricInterface {
+    [key: string]: Array<{
+        exp_id: number,
+        label: string,
+        data: Array<number>,
+        fill: Boolean,
+        backgroundColor: string,
+        borderColor: string,
+        tension: number,
+        pointRadius: Number,
+        pointHitRadius: Number,
+        pointHoverRadius: Number
+    }>
+}
+var lossOrMetricObj: lossOrMetricInterface = {}
+
+const Accordion = styled((props: AccordionProps) => (
+    <MuiAccordion disableGutters elevation={0} square {...props} />))(() => ({
+    '&:not(:last-child)': {
+        borderBottom: 0,
+    },
+    '&:before': {
+        display: 'none',
+    }
+}));
+  
+const AccordionSummary = styled((props: AccordionSummaryProps) => (
+    <MuiAccordionSummary
+      expandIcon={<ArrowForwardIosSharpIcon className={styles.expand_more_icon} />}
+      {...props}
+    />
+  ))(({ theme }) => ({
+    flexDirection: 'row-reverse',
+    '& .MuiAccordionSummary-expandIconWrapper.Mui-expanded': {
+      transform: 'rotate(90deg)',
+    },
+    '& .MuiAccordionSummary-content': {
+      marginLeft: theme.spacing(1),
+    },
+}));
+  
+const AccordionDetails = styled(MuiAccordionDetails)(({ theme }) => ({
+    paddingTop: theme.spacing(0),
+}));
+
+function ExperimentPage() {
+    const [searchParams, setSearchParams] = useSearchParams();
+    let experiment_id = searchParams.get("experiment_id") as string;
+    const [selectedExperiment, setSelectedExperiment] = useState(anyObj)
+    const filteredExp = useAppSelector(state => selectRowById(state, experiment_id));
+    const selectedExpGraphs = useAppSelector(state => selectChartsByExperimentId(state, experiment_id))
+    const isSocketConnected = useAppSelector(isConnected);
+
+    useEffect(() => {
+        console.log("filteredExp: ",filteredExp);
+        console.log("selectedExpGraphs: ",selectedExpGraphs)
+        if (filteredExp) {
+            setSelectedExperiment(filteredExp);
+        }
+    },[filteredExp]);
+
+    return(
+        <div className={styles.main}>
+            <Toolbar />
+            {
+                filteredExp && isSocketConnected ?
+                <>
+                    <Accordion defaultExpanded={true}>
+                        <AccordionSummary>
+                            <div className={styles.accordian_heading}>Details</div>
+                        </AccordionSummary>
+                        <AccordionDetails className={styles.accordian_details}>
+                            <ExperimentDetails selectedExperiment={selectedExperiment}/>
+                        </AccordionDetails>
+                    </Accordion>
+                    <Accordion defaultExpanded={true}>
+                        <AccordionSummary aria-controls="panel2d-content" id="panel2d-header">
+                            <div className={styles.accordian_heading}>Progress</div>
+                        </AccordionSummary>
+                        <AccordionDetails>
+                            <ExperimentProgress selectedExperiment={selectedExperiment}/>
+                        </AccordionDetails>
+                    </Accordion>
+                    {
+                        filteredExp && filteredExp.hasOwnProperty("experiment_id") ?
+                        <Accordion defaultExpanded={true}>
+                            <AccordionSummary aria-controls="panel2d-content" id="panel2d-header">
+                                <div className={styles.accordian_heading}>Configurations</div>
+                            </AccordionSummary>
+                            <AccordionDetails>
+                                <Grid container spacing={{ xs: 2, sm: 2, md: 2, lg: 2 }}>
+                                    <Grid item={true} xs={12} sm={12} md={3}>
+                                        <CheckpointConfigurations experimentIdProp={filteredExp!.experiment_id.toString()}/>
+                                    </Grid>
+                                    <Grid item={true} xs={12} sm={12} md={9}>
+                                        <ExperimentConfigurations experimentIdProp={filteredExp!.experiment_id.toString()}/>
+                                    </Grid>
+                                </Grid>
+                            </AccordionDetails>
+                        </Accordion>
+                        :
+                        null
+                    }
+                    <Accordion defaultExpanded={true}>
+                        <AccordionSummary>
+                            <div className={styles.accordian_heading}>Losses & Metrics</div>
+                        </AccordionSummary>
+                        <AccordionDetails>
+                            {
+                                Object.keys(selectedExpGraphs).length > 0 ?
+                                <Grid container rowSpacing={1} spacing={{ xs: 2, md: 3 }} className={styles_graphs.grid_container}>
+                                    {
+                                        Object.keys(selectedExpGraphs).map((chart_id) => 
+                                            <Graph 
+                                                key={chart_id.toString()} 
+                                                chart_id={chart_id.toString()} 
+                                                exp_id={selectedExpGraphs[chart_id].exp_id.toString()}
+                                                exp_data={selectedExpGraphs[chart_id].data}
+                                            />
+                                        )
+                                    }
+                                </Grid>
+                                :
+                                <Box className={styles_graphs.no_data_to_viz}>
+                                    --- No Data To Visualize ---
+                                </Box>
+                            }
+                        </AccordionDetails>
+                    </Accordion>
+                </>
+                :
+                isSocketConnected ?
+                <div className={styles.extra_text_container}>
+                    <div className={styles.error_text_no_data_1}> No Data Available </div>
+                    <div className={styles.error_text_no_data_2}> (Please wait, experiment details might be loading...) </div>
+                </div>
+                :
+                null
+            }
+        </div>
+    )
+}
+
+export default ExperimentPage;
